# services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Seckill Mall Commerce API
  version: 0.4.0
  description: 阶段 4 商城微服务 API。Gateway 显式路由 Identity、Catalog、Cart、Order、Payment 与 Fulfillment；金额字段均使用人民币分。
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: API Gateway
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          $ref: '#/components/responses/TokenPair'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          $ref: '#/components/responses/TokenPair'
        '401':
          $ref: '#/components/responses/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/TokenPair'
        '401':
          $ref: '#/components/responses/Error'
  /addresses:
    get:
      tags: [Address]
      operationId: listAddresses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 地址列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Address'
    post:
      tags: [Address]
      operationId: createAddress
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '201':
          description: 已创建地址
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Address'
  /addresses/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    put:
      tags: [Address]
      operationId: updateAddress
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '200':
          description: 已更新地址
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [Address]
      operationId: deleteAddress
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 已删除地址
        '404':
          $ref: '#/components/responses/Error'
  /products:
    get:
      tags: [Catalog]
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 在售商品分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageEnvelope'
  /products/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      tags: [Catalog]
      operationId: getProduct
      responses:
        '200':
          description: 商品详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductEnvelope'
        '404':
          $ref: '#/components/responses/Error'
  /cart:
    get:
      tags: [Cart]
      operationId: listCartItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 购物车
  /cart/items:
    post:
      tags: [Cart]
      operationId: setCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemInput'
      responses:
        '200':
          description: 购物车商品
  /cart/items/{sku_id}:
    parameters:
      - name: sku_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    delete:
      tags: [Cart]
      operationId: deleteCartItem
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 已删除购物车商品
  /cart/preview:
    get:
      tags: [Cart]
      operationId: previewCheckout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 实时价格与库存结算预览
...
  /orders/{order_id}/receipt:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Order]
      operationId: confirmOrder
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 已确认收货
...
  /orders/{order_id}/shipment:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    get:
      tags: [Fulfillment]
      operationId: getShipment
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 当前用户订单的物流信息
        '404':
          $ref: '#/components/responses/Error'
...
  /admin/orders/{order_id}/ship:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Admin]
      operationId: shipOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [carrier, tracking_no]
              properties:
                carrier:
                  type: string
                  maxLength: 64
                tracking_no:
                  type: string
                  maxLength: 128
      responses:
        '200':
          description: 已发货
        '403':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    OrderID:
      name: order_id
      in: path
      required: true
      schema:
        type: string
...
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    TokenPair:
      description: Access/Refresh Token 对
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TokenPair'
  schemas:
    Envelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        data: {}
        request_id:
          type: string
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 72
    User:
      type: object
      required: [id, email, role, status]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        role:
          type: string
          enum: [customer, admin]
        status:
          type: string
          enum: [active]
    TokenPair:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
        access_expires_at_unix:
          type: integer
          format: int64
        refresh_token:
          type: string
        refresh_expires_at_unix:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
    AddressInput:
      type: object
      required: [recipient, phone, province, city, district, detail]
      properties:
        recipient:
          type: string
          maxLength: 64
        phone:
          type: string
          maxLength: 32
        province:
          type: string
          maxLength: 64
        city:
          type: string
          maxLength: 64
        district:
          type: string
          maxLength: 64
        detail:
          type: string
          maxLength: 255
        is_default:
          type: boolean
    Address:
      allOf:
        - $ref: '#/components/schemas/AddressInput'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
            user_id:
              type: integer
              format: int64
    CartItemInput:
      type: object
      required: [sku_id, quantity]
      properties:
        sku_id:
          type: integer
          format: int64
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 99
...
    SPUInput:
      type: object
      required: [name]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
    SKUInput:
      type: object
      required: [code, name, price_cents, available_stock]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        code:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 255
        price_cents:
          type: integer
          format: int64
          minimum: 0
        available_stock:
          type: integer
          minimum: 0
...
    SKU:
      allOf:
        - $ref: '#/components/schemas/SKUInput'
        - type: object
          required: [id, spu_id, reserved_stock, active]
          properties:
            spu_id:
              type: integer
              format: int64
            reserved_stock:
              type: integer
              readOnly: true
              minimum: 0
            active:
              type: boolean
    Product:
      type: object
      required: [spu, skus, images]
      properties:
        spu:
          allOf:
            - $ref: '#/components/schemas/SPUInput'
            - type: object
              properties:
                category_id:
                  type: integer
                  format: int64
                active:
                  type: boolean
        skus:
          type: array
          items:
            $ref: '#/components/schemas/SKU'
        images:
          type: array
          items:
            type: string
            format: uri
    ProductEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Product'
...
    ProductPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Product'
                offset:
                  type: integer
                limit:
                  type: integer
                total:
                  type: integer